+++ R/explain.R
@@ -22,12 +22,12 @@ explain_column <- function(object, X, column, pred_wrapper, newdata = NULL) {
   p <- ncol(X)  # number of features
   
   # Generate original and sampled feature instances
-  if (is.null(newdata)) {
-    W <- X[sample(n), ]  # shuffle rows
+  if (is.null(newdata)) {  # FIXME: Should sampling be done with replacement?
+    W <- X[sample(n, replace = TRUE), ]  
     O <- genOMat(n, p)
   } else {
-    W <- X[sample(n, size = nrow(newdata)), , drop = FALSE]  # randomly sample rows from full X
-    O <- genOMat(n, p)[sample(n, size = nrow(newdata)), , drop = FALSE]
+    W <- X[sample(n, size = nrow(newdata), replace = TRUE), , drop = FALSE]  # randomly sample rows from full X
+    O <- genOMat(n, p)[sample(n, size = nrow(newdata), replace = TRUE), , drop = FALSE]
     X <- newdata  # observations of interest
   }
   
@@ -68,6 +68,10 @@ explain_column <- function(object, X, column, pred_wrapper, newdata = NULL) {
     
   }
   
+  # Make sure each component of `B` has the same column classes as `X`
+  B[[1L]] <- copy_classes(B[[1L]], y = X)
+  B[[2L]] <- copy_classes(B[[2L]], y = X)
+  
   # Return differences in predictions
   pred_wrapper(object, newdata = B[[1L]]) - 
     pred_wrapper(object, newdata = B[[2L]])  
@@ -79,18 +83,20 @@ explain_column <- function(object, X, column, pred_wrapper, newdata = NULL) {
 #' 
 #' Compute fast (approximate) Shapley values for a set of features.
 #' 
-#' @param object A fitted model object (e.g., a 
-#' \code{\link[randomForest]{randomForest}} object).
+#' @param object A fitted model object (e.g., a \code{\link[ranger]{ranger}} or
+#' an \code{\link[xgboost]{xgboost}} object).
 #'
 #' @param feature_names Character string giving the names of the predictor
 #' variables (i.e., features) of interest. If \code{NULL} (default) they will be
 #' taken from the column names of \code{X}.
 #'
 #' @param X A matrix-like R object (e.g., a data frame or matrix) containing 
-#' ONLY the feature columns from the training data.
+#' ONLY the feature columns from the training data. \strong{NOTE:} This argument
+#' is required whenever \code{exact = FALSE}.
 #'
 #' @param pred_wrapper Prediction function that requires two arguments,
-#' \code{object} and \code{newdata}. The output of this function should be 
+#' \code{object} and \code{newdata}. \strong{NOTE:} This argument is required 
+#' whenever \code{exact = FALSE}. The output of this function should be 
 #' determined according to:
 #'
 #' \describe{
@@ -102,14 +108,29 @@ explain_column <- function(object, X, column, pred_wrapper, newdata = NULL) {
 #' }
 #' 
 #' @param nsim The number of Monte Carlo repetitions to use for estimating each 
-#' Shapley value. Default is 1.
+#' Shapley value (only used when \code{exact = FALSE}). Default is 1. 
+#' \strong{NOTE:} To obtain the most accurate results, \code{nsim} should be set 
+#' as large as feasibly possible.
 #' 
 #' @param newdata A matrix-like R object (e.g., a data frame or matrix) 
-#' containing ONLY the feature columns for the observation(s) of interest. 
-#' Default is \code{NULL} which will produce approximate Shapley values for all 
-#' the rows in \code{X}.
+#' containing ONLY the feature columns for the observation(s) of interest; that 
+#' is, the observation(s) you want to compute explanations for. Default is 
+#' \code{NULL} which will produce approximate Shapley values for all the rows in 
+#' \code{X} (i.e., the training data).
 #' 
-#' @param ... Additional optional arguments to be passed onto
+#' @param adjust Logical indicating whether or not to adjust the sum of the 
+#' estimated Shapley values to satisfy the \emph{additivity} (or 
+#' \emph{local accuracy}) property; that is, to equal the difference between the 
+#' model's prediction for that sample and the average prediction over all the 
+#' training data (i.e., \code{X}).
+#' 
+#' @param exact Logical indicating whether to compute exact Shapley values. 
+#' Currently only available for \code{\link[stats]{lm}} and 
+#' \code{\link[xgboost]{xgboost}} objects. Default is \code{FALSE}. Note 
+#' that setting \code{exact = TRUE} will return explanations for each of the 
+#' \code{\link[stats]{terms}} in an \code{\link[stats]{lm}} object.
+#' 
+#' @param ... Additional optional arguments to be passed on to
 #' \code{\link[plyr]{laply}}.
 #' 
 #' @return A tibble with one column for each feature specified in 
@@ -122,6 +143,13 @@ explain_column <- function(object, X, column, pred_wrapper, newdata = NULL) {
 #' sets) on the \strong{fastshap} GitHub repository: 
 #' \url{https://github.com/bgreenwell/fastshap}.
 #' 
+#' @note Setting \code{exact = TRUE} with a linear model (i.e., an 
+#' \code{\link[stats]{lm}} or \code{\link[stats]{glm}} object) assumes that the
+#' input features are independent. Also, setting \code{adjust = TRUE} is 
+#' experimental and we follow the same approach as in
+#' \href{https://github.com/slundberg/shap}{shap}.
+#' 
+#' @rdname explain
 #' 
 #' @export
 #' 
@@ -134,12 +162,16 @@ explain_column <- function(object, X, column, pred_wrapper, newdata = NULL) {
 #' data(mtcars)
 #' 
 #' # Fit a projection pursuit regression model
-#' mtcars.ppr <- ppr(mpg ~ ., data = mtcars, nterms = 1)
+#' fit <- lm(mpg ~ ., data = mtcars)
 #' 
 #' # Compute approximate Shapley values using 10 Monte Carlo simulations
 #' set.seed(101)  # for reproducibility
-#' shap <- explain(mtcars.ppr, X = subset(mtcars, select = -mpg), nsim = 10, 
+#' shap <- explain(fit, X = subset(mtcars, select = -mpg), nsim = 10, 
 #'                 pred_wrapper = predict)
+#' shap
+#' 
+#' # Compute exact Shapley (i.e., LinearSHAP) values
+#' shap <- explain(fit, exact = TRUE)
 #' shap
 #' 
 #' # Shapley-based plots
@@ -147,40 +179,190 @@ explain_column <- function(object, X, column, pred_wrapper, newdata = NULL) {
 #' autoplot(shap)  # Shapley-based importance plot
 #' autoplot(shap, type = "dependence", feature = "wt", X = mtcars)
 #' autoplot(shap, type = "contribution", row_num = 1)  # explain first row of X
-explain <- function(object, feature_names = NULL, X, nsim = 1, pred_wrapper,
-                     newdata = NULL, ...) {
+explain <- function(object, ...) {
+  UseMethod("explain")
+} 
+
+
+#' @rdname explain
+#' 
+#' @export
+explain.default <- function(object, feature_names = NULL, X = NULL, nsim = 1, 
+                            pred_wrapper = NULL, newdata = NULL, adjust = FALSE,
+                            ...) {
   
-  # TODO:
-  #
-  #   1. Correct sum of SHAP values to equal the predicted output of the model
-  #      minus the average prediction?
-  #
-  #   2. Feature grouping. For assessing the importance of groups of features
-  #      without assuming independence between them; see
-  #      https://github.com/slundberg/ShapleyValues.jl for an implementation in
-  #      Julia.
+  # Experimental patch for more efficiently computing single-row explanations
+  if (!is.null(newdata)) {
+    if (nrow(newdata) == 1L && nsim > 1) {
+      newdata <- newdata[rep(1L, times = nsim), ]  # replicate row `nsim` times
+      res <- explain.default(object, feature_names = feature_names, X = X, 
+                             nsim = 1L, pred_wrapper = pred_wrapper, 
+                             newdata = newdata, adjust = FALSE, ...)
+      phi_avg <- colMeans(res)
+      if (isTRUE(adjust)) {
+        # Compute average training prediction (fnull) and predictions associated
+        # with each explanation (fx)
+        fx <- pred_wrapper(object, newdata = newdata[1L, , drop = FALSE])
+        fnull <- mean(pred_wrapper(object, newdata = X))
+        phi_var <- apply(res, MARGIN = 2, FUN = stats::var)
+        err <- fx - sum(phi_avg) - fnull
+        v <- (phi_var / max(phi_var)) * 1e6
+        adj <- err * (v - (v * sum(v)) / (1 + sum(v)))
+        phi_avg <- phi_avg + adj
+      }
+      res <- tibble::as_tibble(t(phi_avg))
+      class(res) <- c(class(res), "explain")
+      return(res)
+    }
+  }
   
+  # Deal with NULL arguments
   if (is.null(feature_names)) {
     feature_names = colnames(X)
   }
-  res <- plyr::laply(feature_names, .fun = function(x) {
-    reps <- replicate(nsim, {  # replace later with vapply()
-      explain_column(object, X = X, column = x, pred_wrapper = pred_wrapper,
-                     newdata = newdata)  # numeric(nrow(X))
-    })
-    res <- if (is.matrix(reps)) {
-      # apply(reps, MARGIN = 1, FUN = mean)
-      rowMeans(reps)
-    } else {
-      mean.default(reps)
+  if (is.null(X)) {
+    stop("Training features required for approximate Shapley values. Please ",
+         "supply them via the `X` argument; see `?fastshap::explain` for ",
+         "details.", call. = FALSE)
+  }
+  if (is.null(pred_wrapper)) {
+    stop("Prediction function required for approximate Shapley values. Please ",
+         "supply one via the `pred_wrapper` argument; see ",
+         "`?fastshap::explain` for details.", call. = FALSE)
+  }
+  
+  if (isTRUE(adjust)) {  # compute variances and adjust the sum
+    
+    # Need nsim > 1 to compute variances for adjustment
+    if (nsim < 2) {
+      stop("Need `nsim > 1` whenever `adjust = TRUE`.", call. = FALSE)
     }
-  }, ...)
+    
+    # Check for dependencies
+    # if (!requireNamespace("abind", quietly = TRUE)) {
+    #   stop("Package \"abind\" needed for this function to work. Please ",
+    #        "install it.", call. = FALSE)
+    # }
+    # if (!requireNamespace("matrixStats", quietly = TRUE)) {
+    #   stop("Package \"matrixStats\" needed for this function to work. Please ",
+    #        "install it.", call. = FALSE)
+    # }
+    
+    # Compute approximate Shapley values and variances
+    res <- plyr::laply(feature_names, .fun = function(x) {
+      reps <- replicate(nsim, {  # replace later with vapply()
+        explain_column(object, X = X, column = x, pred_wrapper = pred_wrapper,
+                       newdata = newdata)  # numeric(nrow(X))
+      })
+      if (is.matrix(reps)) {
+        abind::abind(rowMeans(reps), matrixStats::rowVars(reps), along = 2)
+      } else {
+        cbind(mean.default(reps), stats::var(reps))
+      }
+    }, ...)
+    
+    # Compute average training prediction (fnull) and predictions associated
+    # with each explanation (fx)
+    if (is.null(newdata)) {  # explain all training rows
+      fnull <- mean(fx <- pred_wrapper(object, newdata = X))
+    } else {  # explain new observation(s)
+      fx <- pred_wrapper(object, newdata = newdata)
+      fnull <- mean(pred_wrapper(object, newdata = X))
+    }
+    
+    # Adjust sum of approximate Shapley values
+    if (length(dim(res)) == 3L) {  # multiple explanations
+      for (i in seq_len(dim(res)[2L])) {
+        err <- fx[i] - sum(res[, i, 1L]) - fnull
+        v <- (res[, i, 2L] / max(res[, i, 2L])) * 1e6
+        adj <- err * (v - (v * sum(v)) / (1 + sum(v)))
+        res[, i, 1L] <- res[, i, 1L] + adj
+      }
+      res <- res[, , 1L]
+    } else {
+      err <- fx - sum(res[, 1L]) - fnull
+      v <- (res[, 2L] / max(res[, 2L])) * 1e6
+      adj <- err * (v - (v * sum(v)) / (1 + sum(v)))
+      res[, 1L] <- res[, 1L] + adj
+      res <- res[, 1L]
+    }
+    
+  } else {  # don't adjust
+    
+    # Compute approximate Shapley values
+    res <- plyr::laply(feature_names, .fun = function(x) {
+      reps <- replicate(nsim, {  # replace later with vapply()
+        explain_column(object, X = X, column = x, pred_wrapper = pred_wrapper,
+                       newdata = newdata)  # numeric(nrow(X))
+      })
+      if (is.matrix(reps)) {
+        rowMeans(reps)
+      } else {
+        mean.default(reps)
+      }
+    }, ...)
+    
+  }
+  
+  # Reformat into a tibble and fix column names
   res <- if (length(feature_names) == 1L) {
     tibble::enframe(res, name = NULL)
   } else {
     res <- tibble::as_tibble(t(res))
   }
   names(res) <- feature_names
+  
+  # Add extra `"explain"` class for printing, plotting, etc.
   class(res) <- c(class(res), "explain")
   res
+
+}
+
+
+#' @rdname explain
+#' 
+#' @export
+explain.lm <- function(object, feature_names = NULL, X, nsim = 1, 
+                       pred_wrapper, newdata = NULL, exact = FALSE, ...) {
+  if (isTRUE(exact)) {  # use LinearSHAP
+    res <- if (is.null(newdata)) {
+      stats::predict(object, type = "terms", ...)
+    } else {
+      stats::predict(object, newdata = newdata, type = "terms", ...)
+    }
+    baseline <- attr(res, which = "constant")  # mean response for all training data
+    res <- tibble::as_tibble(res)
+    attr(res, which = "baseline") <- baseline
+    class(res) <- c(class(res), "explain")
+    res
+  } else {
+    explain.default(object, feature_names = feature_names, X = X, nsim = nsim,
+                    pred_wrapper = pred_wrapper, newdata = newdata, ...)
+  }
+}
+
+
+#' @rdname explain
+#' 
+#' @export
+explain.xgb.Booster <- function(object, feature_names = NULL, X = NULL, nsim = 1, 
+                                pred_wrapper, newdata = NULL, exact = FALSE, 
+                                ...) {
+  if (isTRUE(exact)) {  # use TreeSHAP
+    if (is.null(X) && is.null(newdata)) {
+      stop("Must supply `X` or `newdata` argument (but not both).", 
+           call. = FALSE)
+    }
+    X <- if (is.null(X)) newdata else X
+    res <- stats::predict(object, newdata = X, predcontrib = TRUE, 
+                          approxcontrib = FALSE, ...)
+    res <- tibble::as_tibble(res)
+    attr(res, which = "baseline") <- res[["BIAS"]]
+    res[["BIAS"]] <- NULL
+    class(res) <- c(class(res), "explain")
+    res
+  } else {
+    explain.default(object, feature_names = feature_names, X = X, nsim = nsim,
+                    pred_wrapper = pred_wrapper, newdata = newdata, ...)
+  }
 }